Beginning to port the easy tests
