+++ test/Mono.Linker.Tests/Mono/Linker/Tests/PreserveActionComparisonTests.cs
@@ -1,4 +1,5 @@
-﻿using NUnit.Framework;
+﻿/*
+using NUnit.Framework;
 
 namespace Mono.Linker.Tests
 {
@@ -28,3 +29,4 @@ namespace Mono.Linker.Tests
 		}
 	}
 }
+*/
